+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# 1.1.0
+# 1.1.x
 
 ## Added
 - Support for Authentication Channels Data Stream
 - Properties on Client class to ease access to account information
+- More debug log calls to connection class to provide more indepth information.
 
 ## Fixed
 - Issue where Error codes would cause a reconnect when not necessary